# Encode queries after state checkout