@whatyouhide 2019-05-15
@lexmag and I decided to go with the solution in #148 instead of this one.